sync data

+++ AcademicAdvisor/delete_appointment.php
@@ -1,12 +1,12 @@
 <?php
 require_once 'connection.php';
 
-//if($_SERVER['REQUEST_METHOD']=='POST') {
+if($_SERVER['REQUEST_METHOD']=='POST') {
 	
 
-   // $apointment_id = trim($_POST['appointment_id']) ;
+    $apointment_id = trim($_POST['appointment_id']) ;
   
-   $apointment_id = "5";
+   //$apointment_id = "4";
 
      $sql = "DELETE FROM `appointments` WHERE `appointments`.`appointment_id` = $apointment_id ";
    mysqli_query($conn,$sql);
@@ -15,6 +15,6 @@ require_once 'connection.php';
 		echo "success";		
 
 	
-//}
+ }
 
 ?>
